requirements.txt
python==3.11
pydub
yt-dlp
pyannote.audio
light-the-torch
onnxruntime
yt_dlp
whisper @ git+https://github.com/openai/whisper.git